embedding.py
from dotenv import load_dotenv
import os
from google import genai

# Load API key from .env
load_dotenv()
api_key = os.getenv("GEMINI_API_KEY")
client = genai.Client(api_key=api_key)

# result = client.models.embed_content(
#         model="gemini-embedding-001",
#         contents="What is the meaning of life?")

# print(result.embeddings)

result = client.models.embed_content(
        model="gemini-embedding-001",
        contents= [
            "What is the meaning of life?",
            "What is the purpose of existence?",
            "How do I bake a cake?"
        ])

for embedding in result.embeddings:
    print(embedding)
