requirements.txt
python-dotenv
google-genai
pandas
scikit-learn
